+++ src/environment/_compiled_footer.js
@@ -1,9 +1,0 @@
-// set up for node or AMD
-if (typeof module !== "undefined") {
-	module.exports = App;
-}
-else if (typeof define !== "undefined") {
-	define([], function() {
-		return App;
-	});
-}