+++ src/linkjs/_compiled_header.js
@@ -1,5 +1,0 @@
-// LinkJS
-// ======
-// pfraze 2012
-function noop() {}
-var Link = {};