+++ src/util/_compiled_header.js
@@ -1,0 +1,10 @@
+// Local Utilities
+// ===============
+// pfraze 2013
+
+if (typeof this.local == 'undefined')
+	this.local = {};
+if (typeof this.local.util == 'undefined')
+	this.local.util = {};
+
+(function() {